LayoutTests/media/track/track-webvtt-tc008-timingsnohours.html

<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>            

            var numberTests = 0;
            
            function track1Loaded()
            {
                findMediaElement();
                testExpected("video.textTracks[0].cues.length", "4");
                
                testExpected("video.textTracks[0].cues[0].id","1");
                testExpected("video.textTracks[0].cues[0].startTime","0.0");
                testExpected("video.textTracks[0].cues[0].endTime","30.5");
                testExpected("video.textTracks[0].cues[0].getCueAsSource()", "Bear is Coming!!!!!");
                
                testExpected("video.textTracks[0].cues[1].id","2");
                testExpected("video.textTracks[0].cues[1].startTime","31.0");
                testExpected("video.textTracks[0].cues[1].endTime","60.5");
                testExpected("video.textTracks[0].cues[1].getCueAsSource()", "I said Bear is coming!!!!");

                testExpected("video.textTracks[0].cues[2].id","3");
                testExpected("video.textTracks[0].cues[2].startTime","61.0");
                testExpected("video.textTracks[0].cues[2].endTime","120.5");
                testExpected("video.textTracks[0].cues[2].getCueAsSource()", "I said Bear is coming now!!!!");

                testExpected("video.textTracks[0].cues[3].id","4");
                testExpected("video.textTracks[0].cues[3].startTime","121.0");
                testExpected("video.textTracks[0].cues[3].endTime","180.5");
                testExpected("video.textTracks[0].cues[3].getCueAsSource()", "tab separators");

                allTestsEnded();
            }

            function track2Loaded()
            {
                findMediaElement();
                testExpected("video.textTracks[1].cues.length", "0");

                allTestsEnded();
            }

            function allTestsEnded()
            {
                numberTests++;
                if (numberTests >= 2)
                    endTest();
            }
        </script>
    </head>
    <body>
        <p>Tests cue timings that do not contain hours (they are optional), and tests various syntax errors in timings without hours.</p>
        <video>
            <track src="captions/tc008_timingsnohour.vtt" onload="track1Loaded()">
            <track src="captions/tc008_timingsnohourerrors.vtt" onload="track2Loaded()">
        </video>
    </body>
</html>
