LayoutTests/media/track/track-webvtt-tc008-timingsnohours-expected.txt

Tests cue timings that do not contain hours (they are optional), and tests various syntax errors in timings without hours.

EXPECTED (video.textTracks[0].cues.length == '4') OK
EXPECTED (video.textTracks[0].cues[0].id == '1') OK
EXPECTED (video.textTracks[0].cues[0].startTime == '0.0') OK
EXPECTED (video.textTracks[0].cues[0].endTime == '30.5') OK
EXPECTED (video.textTracks[0].cues[0].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (video.textTracks[0].cues[1].id == '2') OK
EXPECTED (video.textTracks[0].cues[1].startTime == '31.0') OK
EXPECTED (video.textTracks[0].cues[1].endTime == '60.5') OK
EXPECTED (video.textTracks[0].cues[1].getCueAsSource() == 'I said Bear is coming!!!!') OK
EXPECTED (video.textTracks[0].cues[2].id == '3') OK
EXPECTED (video.textTracks[0].cues[2].startTime == '61.0') OK
EXPECTED (video.textTracks[0].cues[2].endTime == '120.5') OK
EXPECTED (video.textTracks[0].cues[2].getCueAsSource() == 'I said Bear is coming now!!!!') OK
EXPECTED (video.textTracks[0].cues[3].id == '4') OK
EXPECTED (video.textTracks[0].cues[3].startTime == '121.0') OK
EXPECTED (video.textTracks[0].cues[3].endTime == '180.5') OK
EXPECTED (video.textTracks[0].cues[3].getCueAsSource() == 'tab separators') OK
EXPECTED (video.textTracks[1].cues.length == '0') OK
END OF TEST
